refactor dataset manager
